chore: remove duplicate getCatalogSkipReason
